refactor(chessboard): consolidate the topological pipeline into one module tree
With seed-and-grow gone, the crate is a single topological pipeline, but
its modules were still split across topological.rs + topological/ +
top-level boosters.rs + cluster/ + pipeline/. Consolidate every pipeline
stage under one `pipeline/` tree with the orchestrator at its root:

  pipeline/
    mod.rs          detect_all_topological / trace_topological + adapter
    inputs.rs       prefilter
    cluster/        axis clustering (mod, assign, slot_coherence)
    boosters.rs     recall boosters
    recover.rs      component recovery + merge
    geometry_check.rs / output.rs / types.rs

Also inline the trivial grow.rs (an 11-line re-export of
projective_grid::shared::grow::GrowResult) at its four use sites.

Pure file-move + module-path restructure (git mv preserves history); the
public API is byte-identical — detect_all_topological / trace_topological
keep their names and re-export from `pipeline`. No logic, constant, or
threshold touched.

Also refreshes stale doc comments in the cluster stage: the slot-coherence
pass was documented against "chess-corners 0.9's DiskFit" and named removed
seed-and-grow stages (BFS / seed / rescue). It is in fact a live recall
safety-net for the still-selectable DiskFit orientation mode (no-op under
RingFit, precision-safe by its bipartite-quality gate); the docs now say so
and reference the topological stages it actually feeds.

Everyday gate green for the crate (build, clippy -D warnings, doc zero
warnings); chessboard + charuco regression gates byte-identical.
