--- crates/calib-targets-chessboard/src/grow.rs
@@ -1,11 +1,0 @@
-//! `grow` stage types shared by the topological dispatch path.
-//!
-//! The pattern-agnostic [`GrowResult`] container — the `(i, j) → index`
-//! labelling plus the cardinal step vectors and rebase parity — lives in
-//! [`projective_grid::shared::grow`]. The chessboard crate re-exports it here
-//! so the topological recovery, boosters, geometry check, and output stages
-//! share one labelled-grid type.
-
-use projective_grid::shared::grow as pg_grow;
-
-pub use pg_grow::GrowResult;
