+++ crates/calib-targets-chessboard/src/lib.rs
@@ -77,22 +77,18 @@
 //! labelling.
 #![deny(missing_docs)]
 
-mod boosters;
 mod circular_stats;
-mod cluster;
 mod corner;
 mod detector;
-mod grow;
 mod mesh_warp;
 mod params;
 mod pipeline;
 mod rectified_view;
-mod topological;
 
 // --- Public contract ---------------------------------------------------
 pub use corner::ChessCorner;
 pub use detector::{ChessboardCorner, ChessboardDetection, Detector};
 pub use mesh_warp::{rectify_mesh_from_grid, MeshWarpError, RectifiedMeshView};
 pub use params::{AdvancedTuning, ChessboardParamsError, DetectorParams, GraphBuildAlgorithm};
+pub use pipeline::{detect_all_topological, trace_topological};
 pub use rectified_view::{rectify_from_chessboard_result, RectifiedBoardView, RectifyError};
-pub use topological::{detect_all_topological, trace_topological};
