+++ crates/calib-targets-chessboard/src/pipeline/cluster/assign.rs
