+++ crates/calib-targets-chessboard/src/pipeline/inputs.rs
