+++ crates/calib-targets-chessboard/src/pipeline/recover.rs
@@ -11,12 +11,12 @@ use calib_targets_core::{GridTransform, GRID_TRANSFORMS_D4};
 use nalgebra::{Point2, Vector2};
 use projective_grid::shared::merge::{merge_components_local, ComponentInput};
 
-use crate::boosters::apply_boosters_with_directional_edge_scale;
-use crate::cluster::{cluster_axes, ClusterCenters};
+use super::boosters::apply_boosters_with_directional_edge_scale;
+use super::cluster::{cluster_axes, ClusterCenters};
 use crate::corner::{CornerAug, CornerStage};
 use crate::detector::{build_detection_from_grow, ChessboardDetection};
-use crate::grow::GrowResult;
 use crate::params::DetectorParams;
+use projective_grid::shared::grow::GrowResult;
 
 pub(super) type LabelledComponent = HashMap<(i32, i32), usize>;
